+++ base/client.jl
@@ -133,7 +133,7 @@ function eval_user_input(ast::ANY, show_value)
 end
 
 syntax_deprecation_warnings(warn::Bool) =
-    ccall(:jl_parse_depwarn, Cint, (Cint,), warn)!=0
+    ccall(:jl_parse_depwarn, Cint, (Cint,), warn) == 1
 
 function syntax_deprecation_warnings(f::Function, warn::Bool)
     prev = syntax_deprecation_warnings(warn)
@@ -232,7 +232,8 @@ let reqarg = Set(UTF8String["--home",          "-H",
                             "--output-o",
                             "--output-ji",
                             "--output-bc",
-                            "--bind-to"])
+                            "--bind-to",
+                            "--precompiled"])
     global process_options
     function process_options(opts::JLOptions, args::Vector{UTF8String})
         if !isempty(args)