+++ base/nullable.jl
@@ -27,7 +27,13 @@ end
 
 get(x::Nullable) = x.isnull ? throw(NullException()) : x.value
 
-get{T}(x::Nullable{T}, y) = x.isnull ? convert(T, y) : x.value
+@inline function get{T}(x::Nullable{T}, y)
+    if isbits(T)
+        ifelse(x.isnull, convert(T, y), x.value)
+    else
+        x.isnull ? convert(T, y) : x.value
+    end
+end
 
 isnull(x::Nullable) = x.isnull