+++ base/pcre.jl
@@ -140,4 +140,23 @@ function substring_number_from_name(re, name)
         (Ptr{Void}, Cstring), re, name)
 end
 
+function capture_names(re)
+    name_count = info(re, INFO_NAMECOUNT, UInt32)
+    name_entry_size = info(re, INFO_NAMEENTRYSIZE, UInt32)
+    nametable_ptr = info(re, INFO_NAMETABLE, Ptr{UInt8})
+    names = Dict{Int, ASCIIString}()
+    for i=1:name_count
+        offset = (i-1)*name_entry_size + 1
+        # The capture group index corresponding to name 'i' is stored as a
+        # big-endian 16-bit value.
+        high_byte = UInt16(unsafe_load(nametable_ptr, offset))
+        low_byte = UInt16(unsafe_load(nametable_ptr, offset+1))
+        idx = (high_byte << 8) | low_byte
+        # The capture group name is a null-terminated string located directly
+        # after the index.
+        names[idx] = bytestring(nametable_ptr+offset+1)
+    end
+    names
+end
+
 end # module