+++ test/cmdlineargs.jl
@@ -1,6 +1,6 @@
 # This file is a part of Julia. License is MIT: http://julialang.org/license
 
-let exename = joinpath(JULIA_HOME, Base.julia_exename())
+let exename = `$(joinpath(JULIA_HOME, Base.julia_exename())) --precompiled=yes`
     # --version
     let v = split(readall(`$exename -v`), "julia version ")[end]
         @test Base.VERSION_STRING == chomp(v)
@@ -141,6 +141,14 @@ let exename = joinpath(JULIA_HOME, Base.julia_exename())
     @test readchomp(`$exename --depwarn=no -E "Base.syntax_deprecation_warnings(true)"`) == "false"
     @test readchomp(`$exename --depwarn=yes -E "Base.syntax_deprecation_warnings(false)"`) == "true"
     @test !success(`$exename --depwarn=false`)
+    # test deprecated syntax
+    @test !success(`$exename -e "foo (x::Int) = x * x" --depwarn=error`)
+    # test deprecated method
+    @test !success(`$exename -e "
+        foo() = :foo; bar() = :bar
+        @deprecate foo() bar()
+        foo()
+    " --depwarn=error`)
 
     # --inline
     @test readchomp(`$exename -E "Bool(Base.JLOptions().can_inline)"`) == "true"
@@ -182,4 +190,5 @@ let exename = joinpath(JULIA_HOME, Base.julia_exename())
 
     # issue #10562
     @test readchomp(`$exename -e 'println(ARGS);' ''`) == "UTF8String[\"\"]"
+
 end