+++ test/show.jl
@@ -263,3 +263,9 @@ end
 @test_repr "1 => 2 => 3"
 @test_repr "1 => (2 => 3)"
 @test_repr "(1 => 2) => 3"
+
+# pr 12008
+@test_repr "bitstype A B"
+@test_repr "bitstype 100 B"
+@test repr(:(bitstype A B)) == ":(bitstype A B)"
+@test repr(:(bitstype 100 B)) == ":(bitstype 100 B)"