+++ ui/repl.c
@@ -50,6 +50,7 @@ static const char opts[]  =
 
     // startup options
     " -J, --sysimage <file>     Start up with the given system image file\n"
+    " --precompiled={yes|no}    Use precompiled code from system image if available\n"
     " -H, --home <dir>          Set location of julia executable\n"
     " --startup-file={yes|no}   Load ~/.juliarc.jl\n"
     " -f, --no-startup          Don't load ~/.juliarc (deprecated, use --startup-file=no)\n"
@@ -83,7 +84,7 @@ static const char opts[]  =
     " --math-mode={ieee,fast}   Disallow or enable unsafe floating point optimizations (overrides @fastmath declaration)\n\n"
 
     // error and warning options
-    " --depwarn={yes|no}        Enable or disable syntax and method deprecation warnings\n\n"
+    " --depwarn={yes|no|error}  Enable or disable syntax and method deprecation warnings ('error' turns warnings into errors)\n\n"
 
     // compiler output options
     " --output-o name           Generate an object file (including system image data)\n"
@@ -115,7 +116,8 @@ void parse_opts(int *argcp, char ***argvp)
            opt_bind_to,
            opt_handle_signals,
            opt_output_o,
-           opt_output_ji
+           opt_output_ji,
+           opt_use_precompiled
     };
     static char* shortopts = "+vhqFfH:e:E:P:L:J:C:ip:Ob:";
     static struct option longopts[] = {
@@ -131,6 +133,7 @@ void parse_opts(int *argcp, char ***argvp)
         { "post-boot",       required_argument, 0, 'P' },
         { "load",            required_argument, 0, 'L' },
         { "sysimage",        required_argument, 0, 'J' },
+        { "precompiled",     required_argument, 0, opt_use_precompiled },
         { "cpu-target",      required_argument, 0, 'C' },
         { "procs",           required_argument, 0, 'p' },
         { "machinefile",     required_argument, 0, opt_machinefile },
@@ -207,6 +210,14 @@ void parse_opts(int *argcp, char ***argvp)
         case 'J': // sysimage
             jl_options.image_file = strdup(optarg);
             imagepathspecified = 1;
+            break;
+        case opt_use_precompiled:
+            if (!strcmp(optarg,"yes"))
+                jl_options.use_precompiled = JL_OPTIONS_USE_PRECOMPILED_YES;
+            else if (!strcmp(optarg,"no"))
+                jl_options.use_precompiled = JL_OPTIONS_USE_PRECOMPILED_NO;
+            else
+                jl_errorf("julia: invalid argument to --precompiled={yes|no} (%s)\n", optarg);
             break;
         case 'C': // cpu-target
             jl_options.cpu_target = strdup(optarg);
@@ -325,11 +336,13 @@ void parse_opts(int *argcp, char ***argvp)
             break;
         case opt_depwarn:
             if (!strcmp(optarg,"yes"))
-                jl_options.depwarn = 1;
+                jl_options.depwarn = JL_OPTIONS_DEPWARN_ON;
             else if (!strcmp(optarg,"no"))
-                jl_options.depwarn = 0;
+                jl_options.depwarn = JL_OPTIONS_DEPWARN_OFF;
+            else if (!strcmp(optarg,"error"))
+                jl_options.depwarn = JL_OPTIONS_DEPWARN_ERROR;
             else
-                jl_errorf("julia: invalid argument to --depwarn={yes|no} (%s)\n", optarg);
+                jl_errorf("julia: invalid argument to --depwarn={yes|no|error} (%s)\n", optarg);
             break;
         case opt_inline:
             if (!strcmp(optarg,"yes"))